Rename schema folder to validation
