+++ src/components/ActivityCollector/configValidators.js
@@ -10,13 +10,13 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { string, boolean } from "../../utils/schema";
+import { string, boolean } from "../../utils/validation";
 
 export default {
-    clickCollectionEnabled: boolean().default(true),
-    downloadLinkQualifier: string()
-      .regexp()
-      .default(
-        "\\.(exe|zip|wav|mp3|mov|mpg|avi|wmv|pdf|doc|docx|xls|xlsx|ppt|pptx)$"
-      )
+  clickCollectionEnabled: boolean().default(true),
+  downloadLinkQualifier: string()
+    .regexp()
+    .default(
+      "\\.(exe|zip|wav|mp3|mov|mpg|avi|wmv|pdf|doc|docx|xls|xlsx|ppt|pptx)$"
+    )
 };
