+++ test/unit/helpers/testConfigValidators.js
@@ -9,7 +9,7 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-import { objectOf } from "../../../src/utils/schema";
+import { objectOf } from "../../../src/utils/validation";
 
 export default ({
   configValidators,
