+++ test/unit/specs/utils/validation/createArrayOf.spec.js
@@ -10,48 +10,43 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { number } from "../../../../../src/utils/schema";
+import { arrayOf, string } from "../../../../../src/utils/validation";
 import describeTransformer from "./describeTransformer";
 
-describe("schema::minimum", () => {
+describe("validation::arrayOf", () => {
   describeTransformer(
-    "optional minimum",
-    number()
-      .integer()
-      .minimum(4),
+    "optional array with required values",
+    arrayOf(string().required()),
     [
-      { value: 3, error: true },
-      { value: 4 },
-      { value: 5 },
+      { value: ["foo", undefined], error: true },
+      { value: [true, "bar"], error: true },
+      { value: "non-array", error: true },
+      { value: ["foo"] },
+      { value: ["foo", "bar"] },
+      { value: [] },
       { value: null },
       { value: undefined }
     ]
   );
 
   describeTransformer(
-    "required minimum",
-    number()
-      .integer()
-      .minimum(1)
-      .required(),
+    "optional array with optional values",
+    arrayOf(string().default("hello")),
     [
-      { value: null, error: true },
-      { value: undefined, error: true },
-      { value: 0, error: true },
-      { value: 1 }
+      {
+        value: ["a", null, undefined, "b"],
+        expected: ["a", "hello", "hello", "b"]
+      }
     ]
   );
 
   describeTransformer(
-    "default minimum",
-    number()
-      .integer()
-      .minimum(10)
-      .default(42),
+    "required array with optional values",
+    arrayOf(string()).required(),
     [
-      { value: null, expected: 42 },
-      { value: undefined, expected: 42 },
-      { value: 0, error: true }
+      { value: [null] },
+      { value: null, error: true },
+      { value: undefined, error: true }
     ]
   );
 });
