+++ test/unit/specs/utils/validation/integer.spec.js
@@ -10,36 +10,40 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { string } from "../../../../../src/utils/schema";
+import { number } from "../../../../../src/utils/validation";
 import describeTransformer from "./describeTransformer";
 
-describe("schema::nonEmpty", () => {
-  describeTransformer("optional nonEmpty", string().nonEmpty(), [
-    { value: "key" },
-    { value: "", error: true },
-    { value: null },
-    { value: undefined }
+describe("validation::integer()", () => {
+  describeTransformer("optional integer", number().integer(), [
+    { value: 42.01, error: true },
+    { value: -1.1, error: true },
+    { value: NaN, error: true },
+    { value: 0 },
+    { value: 42 },
+    { value: -1 }
   ]);
+
   describeTransformer(
-    "required nonEmpty",
-    string()
-      .nonEmpty()
+    "required integer",
+    number()
+      .integer()
       .required(),
     [
-      { value: "abc" },
       { value: null, error: true },
-      { value: undefined, error: true }
+      { value: undefined, error: true },
+      { value: 10 }
     ]
   );
+
   describeTransformer(
-    "default nonEmpty",
-    string()
-      .nonEmpty()
-      .default("mydefault"),
+    "default integer",
+    number()
+      .integer()
+      .default(12345),
     [
-      { value: null, expected: "mydefault" },
-      { value: undefined, expected: "mydefault" },
-      { value: "abc" }
+      { value: null, expected: 12345 },
+      { value: undefined, expected: 12345 },
+      { value: 10, expected: 10 }
     ]
   );
 });
